Move AuthStatus file to data dir
